+++ cryptofeed/exchange/bitmex/bitmex_rest.py
@@ -38,7 +38,7 @@ class Bitmex(RestAPI):
 
     def _generate_signature(self, verb: str, url: str, data='') -> dict:
         """
-        verb: GET/POST/PUT
+        verb: GET/POST
         url: api endpoint
         data: body (if present)
         """